pyproject.toml
[project]
name = "dbtt-g5t1"
version = "0.1.0"
description = "Code Repository for IS215 Digital Business (Technologies and Transformation) G5 Team 1"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.4.0",
    "fastapi>=0.129.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.14",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.10",
    "pydantic-settings>=2.13.1",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
]

[dependency-groups]
format=[
    "black[jupyter]==25.1.0",
    "click==8.1.7",
    "isort==6.0.1",
    "nbqa==1.9.1",
    "pre-commit>=4.5.1",
    "ruff==0.11.10",
]
